requirements.txt
openai>=1.13.3
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.32.2
langchain>=0.1.12
langchain-openai>=0.0.8
python-dotenv>=1.0.1
